refactor(phase0): RedisConfig를 modules/redis/config.py 로 이동 (Tidy First)
- consumer/config.py 는 하위호환 re-export 로 유지
- modules/redis/client.py, consumer/message_handler.py import 경로 정리
- PostDailyStatistics 에 (post_id, date DESC) 복합 인덱스 migration 추가
- ConsumerConfig 에 MAX_CONSECUTIVE_ERRORS / HEALTHZ_* env 추가 (Phase 7 준비)
- 행위 불변, 183 tests green
