+++ consumer/config.py
@@ -1,39 +1,30 @@
+"""Consumer 프로세스 전용 설정.
+
+RedisConfig 는 modules.redis.config 로 이동. 여기서는 하위호환을 위해
+re-export 만 유지. 신규 코드는 ``from modules.redis.config import RedisConfig`` 사용 권장.
+"""
+
 import environ
+
+from modules.redis.config import RedisConfig  # noqa: F401  (re-export)
 
 env = environ.Env()
 
 
-class RedisConfig:
-    """Redis configuration for consumer."""
-
-    HOST = env("REDIS_HOST", default="localhost")
-    PORT = env.int("REDIS_PORT", default=6379)
-    PASSWORD = env("REDIS_PASSWORD", default="notion-check-plz")
-    DB = env.int("REDIS_DB", default=0)
-
-    # Queue names
-    QUEUE_STATS_REFRESH = "vd2:queue:stats-refresh"
-    QUEUE_STATS_REFRESH_PROCESSING = "vd2:queue:stats-refresh:processing"
-    QUEUE_STATS_REFRESH_FAILED = "vd2:queue:stats-refresh:failed"
-
-    # Consumer settings
-    BLOCKING_TIMEOUT = 5  # seconds for BRPOP
-    MAX_RETRIES = 3  # maximum retry attempts
-    RETRY_BACKOFF_BASE = 2  # exponential backoff base (seconds)
-
-    # DLQ (Dead Letter Queue) settings
-    # Failed queue 최대 크기 - 초과 시 오래된 메시지부터 삭제
-    # https://redis.io/glossary/redis-queue/
-    MAX_FAILED_QUEUE_SIZE = env.int(
-        "REDIS_MAX_FAILED_QUEUE_SIZE", default=10000
-    )
-
-
 class ConsumerConfig:
-    """Consumer process configuration."""
+    """Consumer 프로세스 공통 설정."""
 
     PROCESS_NAME = "stats-refresh-consumer"
     LOG_LEVEL = env("CONSUMER_LOG_LEVEL", default="INFO")
     GRACEFUL_SHUTDOWN_TIMEOUT = env.int(
         "CONSUMER_GRACEFUL_SHUTDOWN_TIMEOUT", default=30
+    )
+    # Phase 7 준비: 연속 에러 허용치 (5 → 30 상향, env override 가능)
+    MAX_CONSECUTIVE_ERRORS = env.int(
+        "CONSUMER_MAX_CONSECUTIVE_ERRORS", default=30
+    )
+    # Phase 7 /healthz 포트 (내부 bind only)
+    HEALTHZ_PORT = env.int("CONSUMER_HEALTHZ_PORT", default=8081)
+    HEALTHZ_STALE_THRESHOLD_SEC = env.int(
+        "CONSUMER_HEALTHZ_STALE_THRESHOLD_SEC", default=60
     )
